requirements.txt
snowflake-snowpark-python
streamlit
streamlit-extras
pandas